Extract putMatch into its own file
and comment out transactions. Has been positively
tested manually. I.e., at least something works.
