+++ backend/server/test/match-curl-tests.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+# A quick file for manual testing
+
 BASE_URL=http://localhost:8080/api/match
 
 # -v
